Sum of minor changes and bug fixes with errors and file refactoring.
