+++ .gitignore
@@ -5,7 +5,7 @@ Data/*
 *.fits
 *.zip
 
-# Nor do VS Community configuration files.
+# Nor do VS Community or VS code configuration files.
 *.suo
 .suo
