+++ core/configuration.py
@@ -196,3 +196,114 @@ def write_configuration_file(config_file_name, config_object,
     config.write()
 
     return None
+
+def copy_configuration_file(config_type, destination, file_name=None):
+    """ This function finds a configuration file based on the 
+    configuration type, and copies it directly into the destination
+    directory. If the configuration file doesn't not exist within 
+    the library, an exception is raised.
+    
+    Simple substring testing is used to check if the configuration
+    type matches.
+
+    Parameters
+    ----------
+    config_type : string
+        The type of configuration file that can be copied.
+    destination : string
+        The destination directory for the configuration file type
+        copy.
+    file_name : string (optional)
+        This is an optional file name. The extension .ini will be 
+        added unless it is already present. No directory information
+        should be provided here.
+
+    Returns
+    -------
+    config_path : string
+        The new path that the configuration can be found at.
+    """
+
+    # Obtaining the list of configuration files.
+    config_files = core.runtime.get_configuration_files()
+    # This is mostly for error messages and warnings as it is a 
+    # nice string of the available configurations.
+    avaliable_config_types = ', '.join(
+        [keydex for keydex, __ in config_files.items()])
+
+    # Checking for valid configuration files.
+    matching_files = {}
+    for keydex, pathdex in copy.deepcopy(config_files).items():
+        if (config_type in keydex):
+            # The configuration file matches 
+            matching_files[keydex] = pathdex
+        else:
+            # It is not a matching file.
+            continue
+
+    # There should only be one matching file. If there is an improper
+    # number, then inform the user.
+    if (len(matching_files) == 0):
+        # There are no matching configurations.
+        raise core.error.InputError("No configuration type `{type}` "
+                                    "matches found. Nothing can be copied. "
+                                    "The available configurations are: "
+                                    "\n    {available}"
+                                    .format(type=config_type, 
+                                            available=avaliable_config_types))
+    elif (len(matching_files) == 1):
+        # All is normal, the configuration file should be copied 
+        # into the new directory, the file name is added too.
+        # Check on if the directory provided is a valid one.
+        if (os.path.isdir(destination)):
+            dir = destination
+        elif (os.path.isfile(destination)):
+            # Attempt to extract only the needed directory 
+            # information.
+            dir, __, __ = core.strformat.split_pathname(pathname=destination)
+        else:
+            raise core.error.InputError("The destination provided is not a "
+                                        "valid directory. Destination: "
+                                        "`{dest}`"
+                                        .format(dest=destination))
+        # The source path for the copied directory file. This should
+        # be okay as there is only one entry in the dictionary.
+        source_path = list(matching_files.values())[0]
+        # Constructing the destination path name for the copied file.
+        __, file_name, __ = core.strformat.split_pathname(pathname=file_name)
+        config_path = core.strformat.combine_pathname(
+            directory=dir, file_name=file_name, extension='.ini')
+
+        # Inform that the file is being copied.
+        core.error.ifas_info("The configuration file matching {type} is "
+                             "being copied. \n "
+                             "Source: {src}   Destination: {dest}"
+                             .format(type=config_type, src=source_path, 
+                                     dest=config_path))
+        # Copying the file.
+        shutil.copyfile(source_path, config_path, follow_symlinks=True)
+
+        # Returning the path in the event that they need it. 
+        return config_path
+    elif (len(matching_files) >= 2):
+        # It is indeterminable as to which file should be coped.
+        raise core.error.InputError("The configuration type `{type}` "
+                                    "returned more than one valid "
+                                    "configuration file. Narrow your type "
+                                    "using more exact inputs for the "
+                                    "available configuration files."
+                                    "\n Matching:  {match} "
+                                    "\n Available:  {available}"
+                                    .format(type=config_type,
+                                            match=', '.join(
+                                                [keydex for keydex, __ 
+                                                 in matching_files.items()]),
+                                            available=avaliable_config_types))
+    else:
+        # There is no reason why the code should enter here as the 
+        # lengths should be defined.
+        raise core.error.BrokenLogicError
+
+    # All done. Though, it should still not enter here either.
+    raise core.error.BrokenLogicError
+    return None
