+++ core/configuration_files/masking_configuration.ini
@@ -52,9 +52,6 @@ filter_tag_name = ''
     # masked.
     run_mask_everything = False
 
-# Working parameters for validation based masks.
-[validate]
-
 
 # Working parameters for value based filters.
 [filter]
