+++ core/configuration_files/tutorial_configuration.ini
@@ -18,8 +18,8 @@ tutorial_creation_override = True
     generation_mode = ''
     # Depending on the generation mode, these parameters may also be
     # needed.
-    # If `fill` is used for the generation mode, all data numbers will
-    # be this value.
+    # If `fill` is used for the generation mode, all data numbers 
+    # will be this value.
     fill_value = 0
     # If `pseudorandom` is used for the generation mode, the random
     # number generator will use this as the seed.
