+++ special.py
@@ -1,0 +1,21 @@
+
+"""
+This is a special file. It contains many script functions which
+do not fit the normal work-flow of this package.
+"""
+
+def script_special_create_tutorial_configuration_file(config):
+    """ This function copies a tutorial configuration file into
+    a target directory. It serves to initiate the tutorials 
+    provided.
+
+    Parameters
+    ----------
+    config : ConfigObj
+        The configuration object that is to be used for this 
+        function.
+
+    Returns
+    -------
+    None    
+    """
