+++ tutorial/__init__.py
@@ -4,4 +4,5 @@ This is the section of the code-base that is more or less geared
 towards the tutorial for the user.
 """
 
+from IfA_Smeargle.tutorial import copying as copy
 from IfA_Smeargle.tutorial import generation
