+++ tutorial/copying.py
@@ -1,0 +1,7 @@
+
+"""
+This module deals with the copying of files from one location to 
+another. This is usually used for building the tutorial.
+"""
+
+# Not used for now.
